changed mask float type

+++ micro_dl/input/inference_dataset.py
@@ -186,7 +186,7 @@ class InferenceDataSet(keras.utils.Sequence):
             im_stack = np.stack(im_stack, axis=self.n_dims - 2)
         # Make sure image has the right dtype if mask (not normalized)
         if not normalize:
-            im_stack = im_stack.astype(np.float64)
+            im_stack = im_stack.astype(np.float32)
         return im_stack
 
     def __getitem__(self, index):
